3rd commit - testested locally re CarDetails.
